@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.7.2
+  sdk: ^3.9.2
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -40,16 +40,24 @@ dependencies:
   dio: ^5.9.0
   get_it: ^8.2.0
   injectable: ^2.5.1
+  retrofit: ^4.6.0
+  logger: ^2.6.0  # for logging purpose
+  json_annotation: ^4.9.0
+  auto_mappr_annotation: ^2.3.0
+  flutter_secure_storage: ^9.2.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^5.0.0
+  flutter_lints: ^6.0.0
 
-  build_runner: ^2.4.15
-  easy_localization_generator: ^0.3.0
+  build_runner: ^2.7.1
   injectable_generator: ^2.6.2
+
+  retrofit_generator: ^10.0.1
+  json_serializable: ^6.11.1
+  auto_mappr: ^2.10.0
 
 flutter:
 
