@@ -3,16 +3,16 @@ import 'package:flutter/material.dart';
 class ButtonBase extends StatelessWidget {
   const ButtonBase({
     super.key,
-    required this.onPressed,
-    required this.child,
-    this.style,
-  });
+    required void Function() onPressed,
+    required Widget child,
+    ButtonStyle? style,
+  }) : _style = style, _child = child, _onPressed = onPressed;
 
-  final VoidCallback onPressed;
+  final VoidCallback _onPressed;
 
-  final Widget child;
+  final Widget _child;
 
-  final ButtonStyle? style;
+  final ButtonStyle? _style;
 
   @override
   Widget build(BuildContext context) {
@@ -20,13 +20,13 @@ class ButtonBase extends StatelessWidget {
     final ColorScheme colorScheme = themeData.colorScheme;
 
     return ElevatedButton(
-      onPressed: onPressed,
+      onPressed: _onPressed,
       style: ElevatedButton.styleFrom(
         elevation: 0,
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
         backgroundColor: colorScheme.primary,
-      ).merge(style),
-      child: child,
+      ).merge(_style),
+      child: _child,
     );
   }
 }
